[tree-view] Move the Rich-only addItems tests to their own file
Split the addItems suite out of the shared TreeViewItemsPlugin tests so the Rich vs Simple division is explicit: the new file is typed with the Rich store and skips the Simple variant, while the shared suite uses the common store and no longer exposes addItems on the Simple apiRef.
