+++ test/utils/tree-view/describeTreeView/index.ts
@@ -1,6 +1,5 @@
 export { describeTreeView } from './describeTreeView';
 export type {
-  DescribeTreeViewRenderer,
   DescribeTreeViewRendererReturnValue,
   DescribeTreeViewRendererUtils,
 } from './describeTreeView.types';
